set the zoom level back to 18 on the GetNearbyPlaceData
added bool to only update location when I tell it to.
example - when the button is clicked to start a physical activity then the location will need to be updated often.

when the app first starts it just needs to get the location and not update again unit told to do so, bu the find my location button.
this way you can look for nearby parks or what ever without your location being constantly updated.